a lot of stuff :thinking
