+++ tests/common_fixtures.py
@@ -96,7 +96,6 @@ def config_test_base():
     mlrun.api.utils.singletons.k8s._k8s = None
     mlrun.api.utils.singletons.logs_dir.logs_dir = None
 
-    mlrun.k8s_utils._k8s = None
     mlrun.runtimes.runtime_handler_instances_cache = {}
     mlrun.runtimes.utils.cached_mpijob_crd_version = None
     mlrun.runtimes.utils.cached_nuclio_version = None
